Massive API documentation update and fixes. This is still a work in progress and will take somewhere between few weeks to few months before it is in a decent state. Any new code MUST adhere to doxygen documentation standards before it can be accepted in KStars.
CCMAIL:kstars-devel@kde.org

+++ kstars/skycomponents/starblocklist.cpp
@@ -89,7 +89,7 @@ bool StarBlockList::fillToMag( float maglim ) {
     if( readOffset <= 0 )
         readOffset = dSReader->getOffset( trixelId );
 
-    Q_ASSERT( nBlocks == blocks.size() );
+    Q_ASSERT( nBlocks == (unsigned int) blocks.size() );
 
     BinFileHelper::unsigned_KDE_fseek( dataFile, readOffset, SEEK_SET );
